// src/shared/proto/host/window.ts
export interface ShowOpenDialogueRequest {
  canSelectMany?: boolean | undefined;
  openLabel?: string | undefined;
  filters?: ShowOpenDialogueFilterOption | undefined;
}

export interface ShowOpenDialogueFilterOption {
  files: string[];
}

export interface SelectedResources {
  paths: string[];
}
// ...
export const ShowOpenDialogueRequest: MessageFns<ShowOpenDialogueRequest> = {
  encode(message: ShowOpenDialogueRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.canSelectMany !== undefined) {
      writer.uint32(16).bool(message.canSelectMany);
    }
    if (message.openLabel !== undefined) {
      writer.uint32(26).string(message.openLabel);
    }
    if (message.filters !== undefined) {
      ShowOpenDialogueFilterOption.encode(message.filters, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ShowOpenDialogueRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseShowOpenDialogueRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.canSelectMany = reader.bool();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.openLabel = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.filters = ShowOpenDialogueFilterOption.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ShowOpenDialogueRequest {
    return {
      canSelectMany: isSet(object.canSelectMany) ? globalThis.Boolean(object.canSelectMany) : undefined,
      openLabel: isSet(object.openLabel) ? globalThis.String(object.openLabel) : undefined,
      filters: isSet(object.filters) ? ShowOpenDialogueFilterOption.fromJSON(object.filters) : undefined,
    };
  },

  toJSON(message: ShowOpenDialogueRequest): unknown {
    const obj: any = {};
    if (message.canSelectMany !== undefined) {
      obj.canSelectMany = message.canSelectMany;
    }
    if (message.openLabel !== undefined) {
      obj.openLabel = message.openLabel;
    }
    if (message.filters !== undefined) {
      obj.filters = ShowOpenDialogueFilterOption.toJSON(message.filters);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ShowOpenDialogueRequest>, I>>(base?: I): ShowOpenDialogueRequest {
    return ShowOpenDialogueRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ShowOpenDialogueRequest>, I>>(object: I): ShowOpenDialogueRequest {
    const message = createBaseShowOpenDialogueRequest();
    message.canSelectMany = object.canSelectMany ?? undefined;
    message.openLabel = object.openLabel ?? undefined;
    message.filters = (object.filters !== undefined && object.filters !== null)
      ? ShowOpenDialogueFilterOption.fromPartial(object.filters)
      : undefined;
    return message;
  },
};

function createBaseShowOpenDialogueFilterOption(): ShowOpenDialogueFilterOption {
  return { files: [] };
}

export const ShowOpenDialogueFilterOption: MessageFns<ShowOpenDialogueFilterOption> = {
  encode(message: ShowOpenDialogueFilterOption, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.files) {
      writer.uint32(10).string(v!);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ShowOpenDialogueFilterOption {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseShowOpenDialogueFilterOption();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.files.push(reader.string());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ShowOpenDialogueFilterOption {
    return { files: globalThis.Array.isArray(object?.files) ? object.files.map((e: any) => globalThis.String(e)) : [] };
  },

  toJSON(message: ShowOpenDialogueFilterOption): unknown {
    const obj: any = {};
    if (message.files?.length) {
      obj.files = message.files;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ShowOpenDialogueFilterOption>, I>>(base?: I): ShowOpenDialogueFilterOption {
    return ShowOpenDialogueFilterOption.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ShowOpenDialogueFilterOption>, I>>(object: I): ShowOpenDialogueFilterOption {
    const message = createBaseShowOpenDialogueFilterOption();
    message.files = object.files?.map((e) => e) || [];
    return message;
  },
};

function createBaseSelectedResources(): SelectedResources {
  return { paths: [] };
}

export const SelectedResources: MessageFns<SelectedResources> = {
  encode(message: SelectedResources, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.paths) {
      writer.uint32(10).string(v!);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SelectedResources {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSelectedResources();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.paths.push(reader.string());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SelectedResources {
    return { paths: globalThis.Array.isArray(object?.paths) ? object.paths.map((e: any) => globalThis.String(e)) : [] };
  },

  toJSON(message: SelectedResources): unknown {
    const obj: any = {};
    if (message.paths?.length) {
      obj.paths = message.paths;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SelectedResources>, I>>(base?: I): SelectedResources {
    return SelectedResources.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SelectedResources>, I>>(object: I): SelectedResources {
    const message = createBaseSelectedResources();
    message.paths = object.paths?.map((e) => e) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isObject(value: any): boolean {
  return typeof value === "object" && value !== null;
}

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}